Fix ordering of CORS filter
The previous ordering of filters in LifligBasicApiSetup.create caused
CORS errors in one of our frontend applications talking to a backend
using this library. The error occurred since an endpoint threw an
exception, triggering the CatchUnhandledThrowablesFilter in this
library. But since that filter was applied _before_ the CORS filter,
the returned error response would not include the correct CORS header.
Thus, the frontend application was unable to show the proper error
response.

This commit moves the CORS filter ahead of all other filters, except the
LastResortCatchAllThrowablesFilter, as we really want that to always
catch things (and we're not that concerned with CORS errors in that
case, as that should be caught in the CatchUnhandledThrowablesFilter
before anyways).