Add ErrorResponseBody.toResponse utility function
In the Tomra project, we have a service that calls another service,
where we want to forward the response in case of an error. It's useful
here to turn an ErrorResponseBody back into a Response, and we figured
that this is a utility function that may be useful to others as well.