Add functions to exclude request/response bodies from logs
We encountered a use case where we want to have `logHttpBody = true` for
the API as a whole, but exclude the response body for a single endpoint.
This commit adds two new functions, `excludeRequestBodyFromLog` and
`excludeResponseBodyFromLog`, that use the request context to exclude
bodies from logs by the LoggingFilter for a given request.