Make http4k Filter objects into classes
At the moment, all the http4k Filters in this library are objects with
an invoke() method, which acts like a constructor in every way. I
personally found this confusing when first browsing the library, and I
think it makes the code less clear (we have invoke() that returns a
Filter which takes a lambda which returns another lambda...).

Instead, I think we should use classes here, that implement the Filter
interface. After all, what we want to do is let users construct filters,
and that is what a class does.