Fix exceptions not being included in logs when using errorResponse
The `errorResponse` function in this library calls
`Request.attachThrowableToLog`, which attaches the exception to the
global `errorLogLens`. This is supposed to be used by the
`LoggingFilter` to log exceptions.

However, we created another `errorLogLens` in `LifligBasicApiSetup`,
which is what was actually passed to the `LoggingFilter`. This meant
that the global `errorLogLens` was never actually used, and any
exceptions passed to `errorResponse` would never be logged.

This commit fixes that.