Fix reading of request/response bodies in LoggingFilter
We encountered a bug in production where request bodies would not be
logged. It turned out that this was due to how we read bodies in the
LoggingFilter here. Our tests did not cover this previously, since they
just tested http4k in memory, whereas using an actual HTTP server
requires different handling in some cases.