Support setting custom log level on LoggingFilter request log
Previously, `LoggingFilter` logged at the `ERROR` log level for 500
responses, and `INFO` for everything else. This mixes 2 concerns:
- The status code to return to the client
- The log level to use for the internal log

Mixing these two concerns has led to some unfortunate behavior from some
library users: changing the status code for a given response to 500, in
order to get the side effect of a more severe request log, when a
different status code would be more appropriate.

It would be better to allow library users to override the default log
level for the request log. This commit does this, by adding a new
`requestLogLevel` flag to `RequestContext`, and allowing users to set
this through the new `severity` parameter on `errorResponse`, and the
`Request.setLogLevel` extension function.