Add attachPrincipalLog and getPrincipalLog request extension functions
In previous versions of http4k, one could set a RequestContextKey lens
on a request, and have that context be available in _outer_ filters of
the http4k filter stack. That changed with http4k v6.

In some of our projects, we relied on the previous mechanism in order to
extract the PrincipalLog of a request for logging. That no longer works.
Elsewhere in this library, we have solved this with a new RequestContext
abstraction. We now extend our RequestContext to allow attaching a
PrincipalLog to it, so that users of the library can attach principals
in their auth filters, and extract them later for logging.